// src/apis/models/FireBytes_Unified_Wave_DeviceMessagePanels_Dtos_GetDeviceMessagePanelsInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_Wave_DeviceMessagePanels_Dtos_MessagePanelTableFilterDto } from './FireBytes_Unified_Wave_DeviceMessagePanels_Dtos_MessagePanelTableFilterDto';
export type FireBytes_Unified_Wave_DeviceMessagePanels_Dtos_GetDeviceMessagePanelsInput = {
    maxResultCount?: number;
    skipCount?: number;
    filter?: FireBytes_Unified_Wave_DeviceMessagePanels_Dtos_MessagePanelTableFilterDto;
    sorting?: string | null;
};